modificações no botao de salvar e cancelar, no contador e no login
